requirements.txt
fastapi
uvicorn
routeros-api
sqlalchemy
cryptography
jinja2
apscheduler
python-multipart
python-dotenv